fix: harden all filesystem drivers, carver formats, and the web API
- fs/: squashfs block_size validation (kills SIGFPE on hostile images);
  minix ninodes geometry cap; UFS ncg/ipg bounds and itable-offset check;
  JFS dinode cap; romfs dataOff u32->u64; btrfs chunk-map validation at
  both parse sites, subtraction-form physical mapping, inline ramBytes
  clamp, disk-bytenr wrap guard; FAT orphan-chain cap; optical claimed
  span clamp; xfs inobt unsigned math + agblocks volume bound + alloc
  clamp; apple physBlock guard; f2fs alloc clamp
- fmt/: overflow guards in avro/macho/pickle/qed/xz/zstd/bigtiff; util
  base64 decode thread-safe
- server: session token for every instance (closes local LPE via the
  elevate handover token and CSRF shutdown; non-root fails open with a
  warning, root fails closed); snapshot file refs under the store lock in
  content/hex/preview (UAF race against live deep jobs); param clamps
  (block_size, retry_passes, raid members, extract indices); mutex on the
  handover token lifecycle (elevate/status, handover, waitpid cleanup);
  presence websocket verifies the token (WS bypasses the pre-routing gate)
  and caps sockets; fix null deref in the hex handler (crash on
  job=missing)
- tests: verify.sh authenticates its web-API checks with the session token

+++ src/carve/fmt/xz.cpp
@@ -19,15 +19,17 @@ namespace ghost {
 // byte carries the least significant bits). Returns the byte length
 // consumed and stores the value; 0 on malformed input.
 static int xzVint(const u8* b, size_t n, i64* out) {
-    i64 v = 0;
+    u64 v = 0;
     size_t i = 0;
     for (; i < n && i < 9; i++) {
         if (i == 8) return 0;
-        v |= (i64)(b[i] & 0x7F) << (7 * i);
+        // Unsigned accumulation: the top 7-bit group can reach shift 49,
+        // which would be signed-overflow UB on a hostile byte.
+        v |= (u64)(b[i] & 0x7F) << (7 * i);
         if ((b[i] & 0x80) == 0) break;
     }
     if (i >= n || (b[i] & 0x80) != 0) return 0;
-    *out = v;
+    *out = (i64)v;
     return (int)i + 1;
 }
